+++ src/datasources/data-frame/datasources/v2/DataFrameDataSourceV2.ts
@@ -1,4 +1,4 @@
-import { AppEvents, DataFrameDTO, DataQueryRequest, DataSourceInstanceSettings, FieldType, LegacyMetricFindQueryOptions, MetricFindValue, TimeRange } from "@grafana/data";
+import { AppEvents, DataFrameDTO, DataQueryRequest, DataSourceInstanceSettings, FieldType, LegacyMetricFindQueryOptions, MetricFindValue, ScopedVars, TimeRange } from "@grafana/data";
 import { DataFrameDataSourceBase } from "../../DataFrameDataSourceBase";
 import { BackendSrv, getBackendSrv, TemplateSrv, getTemplateSrv } from "@grafana/runtime";
 import { Column, Option, DataFrameDataQuery, DataFrameDataSourceOptions, DataFrameQueryType, DataFrameQueryV2, DataFrameVariableQuery, DataFrameVariableQueryType, DataTableProjectionLabelLookup, DataTableProjections, DataTableProperties, defaultQueryV2, defaultVariableQueryV2, FlattenedTableProperties, TableDataRows, TableProperties, TablePropertiesList, ValidDataFrameQuery, ValidDataFrameQueryV2, ValidDataFrameVariableQuery, DataFrameQueryV1 } from "../../types";
@@ -12,6 +12,7 @@ import { catchError, combineLatestWith, from, lastValueFrom, map, Observable, of
 
 export class DataFrameDataSourceV2 extends DataFrameDataSourceBase<DataFrameQueryV2> {
     defaultQuery = defaultQueryV2;
+    private scopedVars: ScopedVars = {};
 
     public constructor(
         public readonly instanceSettings: DataSourceInstanceSettings<DataFrameDataSourceOptions>,
@@ -25,12 +26,13 @@ export class DataFrameDataSourceV2 extends DataFrameDataSourceBase<DataFrameQuer
         query: DataFrameQueryV2,
         options: DataQueryRequest<DataFrameQueryV2>
     ): Observable<DataFrameDTO> {
+        this.scopedVars = options.scopedVars;
         const processedQuery = this.processQuery(query);
 
         if (processedQuery.dataTableFilter) {
-            processedQuery.dataTableFilter = transformComputedFieldsQuery(
-                this.templateSrv.replace(processedQuery.dataTableFilter, options.scopedVars),
-                this.dataTableComputedDataFields,
+            processedQuery.dataTableFilter = this.transformQuery(
+                processedQuery.dataTableFilter,
+                options.scopedVars
             );
         }
 
@@ -52,9 +54,9 @@ export class DataFrameDataSourceV2 extends DataFrameDataSourceBase<DataFrameQuer
         const processedQuery = this.processVariableQuery(query);
 
         if (processedQuery.dataTableFilter) {
-            processedQuery.dataTableFilter = transformComputedFieldsQuery(
-                this.templateSrv.replace(processedQuery.dataTableFilter, options.scopedVars),
-                this.dataTableComputedDataFields,
+            processedQuery.dataTableFilter = this.transformQuery(
+                processedQuery.dataTableFilter,
+                options?.scopedVars! || ''
             );
         }
 
@@ -82,7 +84,7 @@ export class DataFrameDataSourceV2 extends DataFrameDataSourceBase<DataFrameQuer
     shouldRunQuery(query: ValidDataFrameQuery): boolean {
         const processedQuery = this.processQuery(query);
 
-        return !processedQuery.hide && processedQuery.type === DataFrameQueryType.Properties;
+        return !processedQuery.hide;
     }
 
     processQuery(query: DataFrameDataQuery): ValidDataFrameQueryV2 {
@@ -130,6 +132,7 @@ export class DataFrameDataSourceV2 extends DataFrameDataSourceBase<DataFrameQuer
             const dataTableFilter = tableId ? `id = "${tableId}"` : '';
 
             return {
+                ...defaultVariableQueryV2,
                 queryType: DataFrameVariableQueryType.ListColumns,
                 dataTableFilter,
                 ...baseQueryProps,
@@ -189,7 +192,11 @@ export class DataFrameDataSourceV2 extends DataFrameDataSourceBase<DataFrameQuer
     }
 
     public async getColumnOptions(filter: string): Promise<Option[]> {
-        const tables = await lastValueFrom(this.queryTables$(filter, TAKE_LIMIT, [
+        const variableReplacedFilter = this.transformQuery(
+            filter,
+            this.scopedVars
+        );
+        const tables = await lastValueFrom(this.queryTables$(variableReplacedFilter, TAKE_LIMIT, [
             DataTableProjections.ColumnName,
             DataTableProjections.ColumnDataType,
         ]));
@@ -258,7 +265,8 @@ export class DataFrameDataSourceV2 extends DataFrameDataSourceBase<DataFrameQuer
                 });
             }
         });
-        return options;
+        const columnOptionsWithVars = [...this.getVariableOptions(), ...options];
+        return columnOptionsWithVars;
     };
 
     /**
@@ -391,6 +399,13 @@ export class DataFrameDataSourceV2 extends DataFrameDataSourceBase<DataFrameQuer
 
             return value;
         });
+    }
+
+    private transformQuery(query: string, scopedVars: ScopedVars) {
+        return transformComputedFieldsQuery(
+            this.templateSrv.replace(query, scopedVars),
+            this.dataTableComputedDataFields,
+        );
     }
 
     private getFieldsForPropertiesQuery$(processedQuery: ValidDataFrameQueryV2): Observable<DataFrameDTO> {